% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatoR2regex.R
\name{formatoR2regex}
\alias{formatoR2regex}
\title{Convert FormatoR to regex.}
\usage{
formatoR2regex(formatoR)
}
\arguments{
\item{formatoR}{character vector with the formatoR specification for the 
string widths.}
}
\value{
Return a character vector with regular expressions for each component
of the input vector.
}
\description{
\code{formatoR2regex} cnverts FormatoR to regex. FormatoR is a
specification for a fwf schema according to an internal standard in the 
Spanish National Statistics Institute (2020).

This standard specifies character, integer and float types by the letters A,
I and F, respectively. The width of the string is specified with a number. In
the case of float numbers, .d specifies the number of decimal digits after 
the decimal point. See example below
}
\examples{
formatoR <- c('A2', 'I5', 'F5.2')
formatoR2regex(formatoR)


}
