% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwrite_fwf.R
\docType{methods}
\name{fwrite_fwf}
\alias{fwrite_fwf}
\alias{fwrite_fwf,data.frame,character,StfwfSchema-method}
\title{Fast write a fixed-width file.}
\usage{
fwrite_fwf(data, filename, StfwfSchema, validate = FALSE,
  justify = "left", ...)

\S4method{fwrite_fwf}{data.frame,character,StfwfSchema}(data, filename,
  StfwfSchema, validate = FALSE, justify = "right", ...)
}
\arguments{
\item{data}{\linkS4class{data.table} with the data to write.}

\item{filename}{Character vector of length 1 with the name of the file to write.}

\item{StfwfSchema}{Object of class \linkS4class{StfwfSchema} with the schema of the file.}

\item{validate}{Logical vector of length 1 with default value \code{FALSE} to indicate whether to
validate the content of \code{data} before writing.}

\item{justify}{Character vector of length 1 with default value \code{left} to indicate whether to
justify strings to the left or to the right.}

\item{...}{Other parameters from \code{\link[data.table]{fwrite}}.}
}
\value{
Returns an invisible \code{NULL}. The dataset is written in file \code{filename}.
}
\description{
\code{fwrite_fwf} takes as basic input a \linkS4class{data.table} and the schema for
the fwf file to write, concatenates columns accordingly and uses \code{\link[data.table]{fwrite}}
to write the file on disk.
}
\examples{
\dontrun{
# file will be written to working directory
path <- system.file('extdata', package = 'fastReadfwf')
stSchema <- fastReadfwf::xlsxToSchema(file.path(path, 'SchemaSNHS.xlsx'), 'stSchema')
data(MicroDataSNHS)
fwrite_fwf(MicroDataSNHS, file.path(getwd(), 'MicroDataSNHS'), stSchema, justify = 'right')
}

}
\seealso{
\code{\link[data.table]{fwrite}} \code{\link{fread_fwf}]}
}
