% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdf.R
\docType{methods}
\name{getdf}
\alias{getdf}
\alias{getdf,StfwfSchema-method}
\title{Get slot \code{df} from the input object.}
\usage{
getdf(object)

\S4method{getdf}{StfwfSchema}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{StfwfSchema}.}
}
\value{
A data.frame.
}
\description{
\code{getdf} extracts the slot \code{df} of the input object.
}
\examples{
# A trivial example:
df <- data.frame(variable = c('Turnover', 'Employees'),
                 width = c(9L, 3L),
                 initialPos = c(1, 10),
                 finalPos = c(9, 12),
                 type = rep('num', 2),
                 valueRegEx = c('[0-9]{0,9}', '[0-9]{0,3}'),
                 description = c('Turnover of the business unit',
                                 'Number of employees of the business unit'),
                 stringsAsFactors = FALSE)
Schema <- new(Class = 'StfwfSchema', df = df)
getdf(Schema)


}
