% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTypes.R
\docType{methods}
\name{getTypes}
\alias{getTypes}
\alias{getTypes,StfwfSchema-method}
\title{Return variable types from the input schema.}
\usage{
getTypes(object)

\S4method{getTypes}{StfwfSchema}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{StfwfSchema}.}
}
\value{
A character vector with component entries \code{'num'}, \code{'char'}, and/or
\code{'bool'}).
}
\description{
\code{getTypes} extracts the slot \code{df} of the input object and returns its
column \code{type} as a vector.
}
\examples{
# A trivial example:
df <- data.frame(variable = c('Turnover', 'Employees'),
                 length = c(9L, 3L),
                 initialPos = c(1, 10),
                 finalPos = c(9, 12),
                 type = rep('num', 2),
                 valueRegEx = c('[0-9]{0,9}', '[0-9]{0,3}'),
                 description = c('Turnover of the business unit',
                                 'Number of employees of the business unit'),
                 stringsAsFactors = FALSE)
Schema <- new(Class = 'StfwfSchema', df = df)
getTypes(Schema)

}

