% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWidths.R
\name{getWidths}
\alias{getWidths}
\alias{getWidths,StfwfSchema-method}
\title{Return variable lengths from the input schema.}
\usage{
getWidths(object)

\S4method{getWidths}{StfwfSchema}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{StfwfSchema}.}
}
\value{
An integer vector.
}
\description{
\code{getWidths} extracts the slot \code{df} of the input object and returns its
column \code{length} as a vector.
}
\examples{
# A trivial example:
df <- data.frame(variable = c('Turnover', 'Employees'),
                 width = c(9L, 3L),
                 initialPos = c(1, 10),
                 finalPos = c(9, 12),
                 type = rep('num', 2),
                 valueRegEx = c('[0-9]{0,9}', '[0-9]{0,3}'),
                 description = c('Turnover of the business unit',
                                 'Number of employees of the business unit'),
                 stringsAsFactors = FALSE)
Schema <- new(Class = 'StfwfSchema', df = df)
getWidths(Schema)


}
