% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INExmlToSchema.R
\name{INExmlToSchema}
\alias{INExmlToSchema}
\title{Build an object of class \linkS4class{StfwfSchema}.}
\usage{
INExmlToSchema(xmlName)
}
\arguments{
\item{xmlName}{Path of the xml file containing the schema.}
}
\value{
Return an object of class \linkS4class{StfwfSchema}.
}
\description{
\code{xmlToSchema} is a constructor of the class \linkS4class{StfwfSchema}.

This constructor reads an xml file containing totally the schema of the
fixed-width file to read. This file contain a root node named \code{Schema},
a secondary root node named \code{Variables} and one node \code{var} for each
variable of the Schema who has the following child nodes:

\itemize{

   \item \code{variable}: the name of the variable.
   \item \code{width}: the number of positions which the values of
   this variable occupies in the file.
   \item \code{initialPos}: initial position of the field which
   the values of this variable occupies in the file.
   \item \code{finalPos}: final position of the field which the
   values of this variable occupies in the file.
   \item \code{type}: type of the variable. It must be either \code{log},
   \code{integer}, \code{num} or \code{char}.
   \item \code{valueRegEx}: regular expression for the values of
   this variable.
   \item \code{description}: textual description of the variable.

}
}
\examples{

xmlName <- file.path(system.file('inst/extdata', package = 'fastReadfwf'), 'disreg_enceursalud20_a.xml')
xmlToSchema(xmlName)

xmlName <- file.path(system.file('data', package = 'fastReadfwf'), 'dr_EPA_2021.xml')
xmlToSchema(xmlName)


}
