% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setChar.R
\name{setChar}
\alias{setChar}
\title{Set the class of each variable in data to character.}
\usage{
setChar(data)
}
\arguments{
\item{data}{It can be a \linkS4class{data.table} or a \link[tibble]{tibble}.}
}
\value{
Returns a \linkS4class{data.table} or a \link[tibble]{tibble}
with the content of \code{data} but with all the columns set to character.
}
\description{
\code{setChar} Set the class of each variable in \code{data}
to character and the \code{NA} values to empty string.
}
\examples{
path <- system.file('extdata', package = 'fastReadfwf')
stSchema <- fastReadfwf::StxlsxToSchema(file.path(path, 'SchemaSNHS.xlsx'), 'stSchema')

# For data.tables
data.DT.char <- fread_fwf(
     file.path(path, 'MicroDataSNHS.txt'), stSchema, validate = FALSE, convert = FALSE,
       outFormat = 'data.table', perl = TRUE)
data.DT.types <- setTypes(data.DT.char, stSchema)
data.DT.back <- setChar(data.DT.types)
head(data.DT.types)

# For tibbles
 data.tibble.char <- fread_fwf(
      file.path(path, 'MicroDataSNHS.txt'), stSchema, validate = FALSE, convert = FALSE,
       outFormat = 'tibble')
 data.tibble.types <- setTypes(data.tibble.char, stSchema)
 data.tibble.back <- setChar(data.tibble.types)
 head(data.tibble.types)

}
\seealso{
\code{\link[fastReadfwf]{fwrite_fwf}} \code{\link[fastReadfwf]{setTypes}}
}
