% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdf.R
\name{setdf<-}
\alias{setdf<-}
\alias{setdf<-,StfwfSchema,data.frame-method}
\title{Set slot \code{df}.}
\usage{
setdf(object) <- value

\S4method{setdf}{StfwfSchema,data.frame}(object) <- value
}
\arguments{
\item{object}{object with the fwf file schema.}

\item{value}{new data.frame with the specification of the schema.}
}
\value{
object with slot \code{df} updated.
}
\description{
\code{setdf} sets the slot \code{df}.
}
\examples{
# A trivial example:
df <- data.frame(variable = c('Turnover', 'Employees'),
                 width = c(9L, 3L),
                 initialPos = c(1, 10),
                 finalPos = c(9, 12),
                 type = rep('num', 2),
                 valueRegEx = c('[0-9]{0,9}', '[0-9]{0,3}'),
                 description = c('Turnover of the business unit',
                                 'Number of employees of the business unit'),
                 stringsAsFactors = FALSE)
Schema <- new(Class = 'StfwfSchema')
setdf(Schema) <- df

}
