% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setfinalPos.R
\docType{methods}
\name{setfinalPos}
\alias{setfinalPos}
\alias{setfinalPos,StfwfSchema,integer-method}
\title{Set final position(s) of variable(s)}
\usage{
setfinalPos(schema, finalPos, variables)

\S4method{setfinalPos}{StfwfSchema,integer}(schema, finalPos, variables)
}
\arguments{
\item{schema}{Object of class \linkS4class{StfwfSchema}.}

\item{finalPos}{integer vector with the new final positions of the variables.}

\item{variables}{character vector with the names of the variables whose final positions are to be
updated.}
}
\value{
Object of class \linkS4class{StfwfSchema}.
}
\description{
\code{setfinalPos} assigns an integer value (final positition) to the variable(s)
specified in the input parameters.
}
\examples{
# A trivial example:
df <- data.frame(variable = c('Turnover', 'Employees'),
                 width = c(9L, 3L),
                 initialPos = c(1, 10),
                 finalPos = c(9, 12),
                 type = rep('num', 2),
                 valueRegEx = c('[0-9]{0,9}', '[0-9]{0,3}'),
                 description = c('Turnover of the business unit',
                                 'Number of employees of the business unit'),
                 stringsAsFactors = FALSE)
Schema <- new(Class = 'StfwfSchema', df = df)
setfinalPos(Schema, c(8L), c('Turnover'))

}

