% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setinitialPos.R
\docType{methods}
\name{setinitialPos}
\alias{setinitialPos}
\alias{setinitialPos,StfwfSchema,integer-method}
\title{Set initial position(s) of variable(s)}
\usage{
setinitialPos(schema, initialPos, variables)

\S4method{setinitialPos}{StfwfSchema,integer}(schema, initialPos, variables)
}
\arguments{
\item{schema}{Object of class \linkS4class{StfwfSchema}.}

\item{variables}{character vector with the names of the variables whose final positions are to be
updated.}

\item{initial}{integer vector with the new initial positions of the variables.}
}
\value{
Object of class \linkS4class{StfwfSchema}.
}
\description{
\code{setinitialPos} assigns an integer value (initial positition) to the
variable(s) specified in the input parameters.
}
\examples{
# A trivial example:
df <- data.frame(variable = c('Turnover', 'Employees'),
                 width = c(9L, 3L),
                 initialPos = c(1, 10),
                 finalPos = c(9, 12),
                 type = rep('num', 2),
                 valueRegEx = c('[0-9]{0,9}', '[0-9]{0,3}'),
                 description = c('Turnover of the business unit',
                                 'Number of employees of the business unit'),
                 stringsAsFactors = FALSE)
Schema <- new(Class = 'StfwfSchema', df = df)
setinitialPos(Schema, c(2L), c('Turnover'))

}

