% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLengths.R
\docType{methods}
\name{setLengths}
\alias{setLengths}
\alias{setLengths,StfwfSchema,integer-method}
\title{Set length(s) of variable(s)}
\usage{
setLengths(schema, lengths, variables)

\S4method{setLengths}{StfwfSchema,integer}(schema, lengths, variables)
}
\arguments{
\item{schema}{Object of class \linkS4class{StfwfSchema}.}

\item{lengths}{integer vector with the new lengths of the variables.}

\item{variables}{character vector with the names of the variables whose lengths are to be
updated.}
}
\value{
Object of class \linkS4class{StfwfSchema}.
}
\description{
\code{setLengths} assigns an integer value (length) to the variable(s) specified in
the input parameters.
}
\examples{
# A trivial example:
df <- data.frame(variable = c('Turnover', 'Employees'),
                 length = c(9L, 3L),
                 initialPos = c(1, 10),
                 finalPos = c(9, 12),
                 type = rep('num', 2),
                 valueRegEx = c('[0-9]{0,9}', '[0-9]{0,3}'),
                 description = c('Turnover of the business unit',
                                 'Number of employees of the business unit'),
                 stringsAsFactors = FALSE)
Schema <- new(Class = 'StfwfSchema', df = df)
setLengths(Schema, 10, 'Turnover')
setLengths(Schema, c(10L, 4L))

}

