% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setRegEx.R
\name{setRegEx}
\alias{setRegEx}
\alias{setRegEx,StfwfSchema,character-method}
\title{Set regex for the accepted values of the variable(s)}
\usage{
setRegEx(schema, regex, variables)

\S4method{setRegEx}{StfwfSchema,character}(schema, regex, variables)
}
\arguments{
\item{schema}{Object of class \linkS4class{StfwfSchema}.}

\item{regex}{charactec vector with the new regex for the accepted values of the variables.}

\item{variables}{character vector with the names of the variables whose regex are to be
updated.}
}
\value{
Object of class \linkS4class{StfwfSchema}.
}
\description{
\code{setRegEx} assigns a regexp to the values of each variable specified in
the input parameters.
}
\examples{
# A trivial example:
df <- data.frame(variable = c('Turnover', 'Employees'),
                 width = c(9L, 3L),
                 initialPos = c(1, 10),
                 finalPos = c(9, 12),
                 type = rep('num', 2),
                 valueRegEx = c('[0-9]{0,9}', '[0-9]{0,3}'),
                 description = c('Turnover of the business unit',
                                 'Number of employees of the business unit'),
                 stringsAsFactors = FALSE)
Schema <- new(Class = 'StfwfSchema', df = df)
setRegEx(Schema, c('[0-9]{0,10}'), 'Turnover')

}
