% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setWidths.R
\docType{methods}
\name{setWidths}
\alias{setWidths}
\alias{setWidths,StfwfSchema,integer-method}
\title{Set width(s) of variable(s)}
\usage{
setWidths(schema, widths, variables)

\S4method{setWidths}{StfwfSchema,integer}(schema, widths, variables)
}
\arguments{
\item{schema}{Object of class \linkS4class{StfwfSchema}.}

\item{widths}{integer vector with the new widths of the variables.}

\item{variables}{character vector with the names of the variables whose widths are to be
updated.}
}
\value{
Object of class \linkS4class{StfwfSchema}.
}
\description{
\code{setWidths} assigns an integer value (width) to the variable(s) specified in
the input parameters.
}
\examples{
# A trivial example:
df <- data.frame(variable = c('Turnover', 'Employees'),
                 width = c(9L, 3L),
                 initialPos = c(1, 10),
                 finalPos = c(9, 12),
                 type = rep('num', 2),
                 valueRegEx = c('[0-9]{0,9}', '[0-9]{0,3}'),
                 description = c('Turnover of the business unit',
                                 'Number of employees of the business unit'),
                 stringsAsFactors = FALSE)
Schema <- new(Class = 'StfwfSchema', df = df)
setWidths(Schema, c(6L), c('Turnover'))

}

