% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\docType{methods}
\name{show,StfwfSchema-method}
\alias{show,StfwfSchema-method}
\title{Show an object of class \linkS4class{StfwfSchema}.}
\usage{
\S4method{show}{StfwfSchema}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{StfwfSchema}.}
}
\value{
Return invisible \code{\link{NULL}}.
}
\description{
\code{show} displays the slot \code{df} of the input \linkS4class{StfwfSchema} object excluding
the description column.
}
\examples{
# A trivial example
show(new(Class = 'StfwfSchema'))

}

