% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StfwfSchema-class.R
\docType{class}
\name{StfwfSchema-class}
\alias{StfwfSchema-class}
\title{S4 class for the standard data.frame for a fixed-width file schema.}
\description{
Definition of the S4 class named \code{StfwfSchema} for the data.frame providing the
schema of the fixed-width file to read.
}
\section{Slots}{

\describe{
\item{\code{Schemadf}}{data.frame with 5 columns:
 \itemize{

   \item variable: the name of the variable.
   \item length: the number of positions which the values of this variable occupies in the file.
   \item initialPos: initial position of the field whic hthe values of this variable occupies in
   the file.
   \item finalPos: final position of the field whic hthe values of this variable occupies in the
   file.
   \item valueRegExp: regular expression for the values of this variable.
   \item description: textual description of the variable.

 }}
}}

\examples{
# An empty ObsErrorSTDMLEParam object:


}
