% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvToSchema.R
\name{csvToSchema}
\alias{csvToSchema}
\title{Build an object of class \linkS4class{StfwfSchema}.}
\usage{
csvToSchema(csvname, sep = ";", header = FALSE, lang = "en", ...)
}
\arguments{
\item{csvname}{Name of the csv file containing the schema.}

\item{sep}{The separator between columns. Defaults to ';'.}

\item{header}{Does the first data line contain column names? Defaults to \code{FALSE}.}

\item{lang}{Character vector of length 1 indicating the language for the header in the csv file
(English: en -- default). So far only English is supported.}

\item{...}{Extra arguments for \code{\link[data.table]{fread}}.}
}
\value{
Return an object of class \linkS4class{StfwfSchema}.
}
\description{
\code{CSVToSchema} is a constructor of the class \linkS4class{StfwfSchema}.

This constructor reads a csv file containing partially or totally the schema of the fixed-width
file to read. So far, only English names are supported. This file must contain the following
columns:

\itemize{

   \item \code{variable} (en): the name of the variable.
   \item \code{width} (en): the number of positions which the values of each variable occupies in
    the file.
   \item \code{initialPos} (en): initial position of the field which the values of this variable
   occupies in the file.
   \item \code{finalPos} (en): final position of the field which the values of this variable
   occupies in the file.
   \item \code{type} (en): type of the variable. It must be either \code{log}, \code{integer},
   \code{num} or \code{char}.
   \item \code{valueRegEx} (en): regular expression for the values of each variable.
   \item \code{description}: textual description of the variable.

}

The file may have a header or not. In the latter case, the order of columns is assumed to be that
 of the list above. English only is supported so far.
}
\examples{
path <- system.file('extdata', package = 'fastReadfwf')
csvToSchema(file.path(path, 'SchemaSNHS.csv'), header = TRUE)

}
