% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fread_fwf.R
\docType{methods}
\name{fread_fwf}
\alias{fread_fwf}
\alias{fread_fwf,character,StfwfSchema-method}
\title{Fast read a fixed-width file}
\usage{
fread_fwf(filename, StfwfSchema, outFormat, ...)

\S4method{fread_fwf}{character,StfwfSchema}(filename, StfwfSchema,
  outFormat = "data.table", perl = FALSE, ...)
}
\arguments{
\item{filename}{Character vector of length 1 with the name of the file to read.}

\item{StfwfSchema}{Object of class \linkS4class{StfwfSchema} with the schema of the file to read.}

\item{outFormat}{Character vector of length 1 whose default value is "data.table". Other
option is "tibble".}

\item{...}{Other parameters from \code{\link[data.table]{fread}} or \code{\link[readr]{read_fwf}}
according to the value of \code{outFormat} above.}

\item{perl}{Logical vector of length 1 with default value \code{FALSE} to indicate whether to use
perl or not in the application of regexp.}
}
\value{
Returns a \linkS4class{data.table} with the contents of the file.
}
\description{
\code{fread_fwf} takes as basic input a fixed-width filename and its schema and
wraps around \code{\link[data.table]{fread}} from package \linkS4class{data.table} to provide the
contents of the file.

This method is indeed a wrapper for the function \code{\link[data.table]{fread}} followed by the
application of \code{\link[stringi]{stri_sub}}.
}
\examples{
\dontrun{
path <- 'C:/Users/David/Documents/Cursos.Seminarios.Impartidos/UCM/EMOS/Organization/Course 2018-2019/Organization - OOP'
pathSchema <- file.path(path, 'EPA/stEPA2018_Schema.xlsx')
stSchema <- fastReadfwf::XLSToSchema(pathSchema, sheetname = 'stEPA2018_Schema', lang = 'en')
dataFile_T1 <- file.path(path, 'EPA/md_EPA_2018T1.txt')
data_T1_dt <- fread_fwf(dataFile_T1, stSchema, outFormat = 'data.table', perl = TRUE)
data_T1_ti <- fread_fwf(dataFile_T1, stSchema, outFormat = 'tibble')

}
}
\seealso{
\code{\link[data.table]{fread}}
}

