% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fread_fwf.R
\docType{methods}
\name{fread_fwf}
\alias{fread_fwf}
\alias{fread_fwf,character,StfwfSchema-method}
\title{Fast read a fixed-width file}
\usage{
fread_fwf(filename, StfwfSchema, encoding = "unknown", check = FALSE,
  perl = FALSE)

\S4method{fread_fwf}{character,StfwfSchema}(filename, StfwfSchema,
  encoding = "unknown", check = FALSE, perl = FALSE)
}
\arguments{
\item{filename}{Character vector of length 1 with the name of the file to read.}

\item{StfwfSchema}{Object of class \linkS4class{StfwfSchema} with the schema of the file to read.}

\item{encoding}{Character vector of length 1 with default value is "unknown". Other possible
options are "UTF-8" and "Latin-1".
NB: it is not used to re-encode the input but to enable handling of encoded strings in their
native encoding.}

\item{check}{Logical vector of length 1 with default value \code{FALSE} to check if the values of
each variable in the read file satisfy the regexp specified in the schema.}

\item{perl}{Logical vector of length 1 with default value \code{FALSE} to indicate whether to use
perl or not in the application of regexp.}
}
\value{
Returns a \linkS4class{data.table} with the contents of the file.
}
\description{
\code{fread_fwf} takes as basic input a fixed-width filename and its schema and
wraps around \code{\link[data.table]{fread}} from package \linkS4class{data.table} to provide the
contents of the file.

This method is indeed a wrapper for the function \code{\link[data.table]{fread}} followed by the
application of \code{\link[stringi]{stri_sub}}.
}
\examples{
\dontrun{
filename <- 'file1.txt'
fileschema <- XLSToSchema('file1schema.xlsx')
fread_fwf(filename, fileschema)

}
}
\seealso{
\code{\link[data.table]{fread}}
}

