% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fread_fwf.R
\docType{methods}
\name{fread_fwf}
\alias{fread_fwf}
\alias{fread_fwf,character,StfwfSchema-method}
\title{Fast read a fixed-width file}
\usage{
fread_fwf(filename, StfwfSchema, outFormat, ...)

\S4method{fread_fwf}{character,StfwfSchema}(filename, StfwfSchema,
  outFormat = "data.table", perl = FALSE, ...)
}
\arguments{
\item{filename}{Character vector of length 1 with the name of the file to read.}

\item{StfwfSchema}{Object of class \linkS4class{StfwfSchema} with the schema of the file to read.}

\item{outFormat}{Character vector of length 1 whose default value is "data.table". Other
option is "tibble".}

\item{...}{Other parameters from \code{\link[data.table]{fread}} or \code{\link[readr]{read_fwf}}
according to the value of \code{outFormat} above.}

\item{perl}{Logical vector of length 1 with default value \code{FALSE} to indicate whether to use
perl or not in the application of regexp.}
}
\value{
Returns a \linkS4class{data.table} with the contents of the file.
}
\description{
\code{fread_fwf} takes as basic input a fixed-width filename and its schema and
wraps around \code{\link[data.table]{fread}} from package \linkS4class{data.table} to provide the
contents of the file.

This method is indeed a wrapper for the function \code{\link[data.table]{fread}} followed by the
application of \code{\link[stringi]{stri_sub}}.
}
\examples{
path <- system.file('extdata', package = 'fastReadfwf')
stSchema <- fastReadfwf::xlsxToSchema(file.path(path, 'SchemaSNHS.xlsx'), 'stSchema')

# For data.tables
data.DT <- fread_fwf(
file.path(path, 'MicroDataSNHS.txt'), stSchema, outFormat = 'data.table', perl = TRUE)
head(data.DT)

# For tibbles
data.tibble <- fread_fwf(
file.path(path, 'MicroDataSNHS.txt'), stSchema, outFormat = 'tibble')
head(data.tibble)

}
\seealso{
\code{\link[data.table]{fread}}
}
