% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fread_fwf.R
\name{fread_fwf}
\alias{fread_fwf}
\alias{fread_fwf,character,StfwfSchema-method}
\title{Fast read a fixed-width file.}
\usage{
fread_fwf(
  filename,
  StfwfSchema,
  validate = FALSE,
  convert = TRUE,
  outFormat = "data.table",
  perl = FALSE,
  ...
)

\S4method{fread_fwf}{character,StfwfSchema}(
  filename,
  StfwfSchema,
  validate = FALSE,
  convert = TRUE,
  outFormat = "data.table",
  perl = FALSE,
  encoding = "unknown",
  ...
)
}
\arguments{
\item{filename}{Character vector of length 1 with the name of the file to read.}

\item{StfwfSchema}{Object of class \linkS4class{StfwfSchema} with the schema of the file to read.}

\item{validate}{Logical vector of length 1 with default value \code{FALSE} to indicate whether to
validate the content of \code{filename} with the regular expressions in \code{StfwfSchema}. This step is done before
converting the types.}

\item{convert}{Logical vector of length 1 with default value \code{TRUE} to indicate whether to convert
the content of \code{filename} to the types in \code{StfwfSchema}.}

\item{outFormat}{Character vector of length 1 whose default value is "data.table". Other
option is "tibble".}

\item{perl}{Logical vector of length 1 with default value \code{FALSE} to indicate whether to use
perl or not in the application of the column \code{regexp}.}

\item{...}{Other parameters from \code{\link[data.table]{fread}} or \code{\link[readr]{read_fwf}}
according to the value of \code{outFormat} above.}

\item{encoding}{Optional argument with the encoding to be used by fread, by default 'unknown' 
(see \code{\link{fread}}).}
}
\value{
Returns a \linkS4class{data.table} or a \link[tibble]{tibble} with the contents of the
file.
}
\description{
\code{fread_fwf} takes as basic input a fixed-width filename and its schema and
wraps around \code{\link[data.table]{fread}} from package \linkS4class{data.table} or around
\code{\link[readr]{read_fwf}} to provide the contents of the file.

This method is indeed either (i) a wrapper for the function \code{\link[data.table]{fread}}
followed by the application of \code{\link[stringi]{stri_sub}} or (ii) a direct wrapper for the
function \code{\link[readr]{read_fwf}}.
}
\examples{
path <- system.file('extdata', package = 'fastReadfwf')
stSchema <- fastReadfwf::StxlsxToSchema(file.path(path, 'SchemaSNHS.xlsx'), 'stSchema')

# For data.tables
data.DT <- fread_fwf(
file.path(path, 'MicroDataSNHS.txt'), stSchema, outFormat = 'data.table', perl = TRUE)
head(data.DT)

# For tibbles
data.tibble <- fread_fwf(
file.path(path, 'MicroDataSNHS.txt'), stSchema, outFormat = 'tibble')
head(data.tibble)

}
\seealso{
\code{\link[data.table]{fread}}
}
