% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fread_fwf.R
\name{fread_fwf}
\alias{fread_fwf}
\title{Fast read a fixed-width file}
\usage{
fread_fwf(file, schema)
}
\arguments{
\item{schema}{\code{Character} vector with names of the output variables.}

\item{filename}{Object of class \linkS4class{StQ} whose slot \code{Data} will be converted.}

\item{perl}{Logical vector of length 1 indicating whether Perl is installed in the system or not.}

\item{encoding}{Character vector of length 1 with default value is "unknown". Other possible 
options are "UTF-8" and "Latin-1". 
NB: it is not used to re-encode the input but to enable handling of encoded strings in their 
native encoding.}
}
\value{
Returns a \linkS4class{data.table} with the contents of the file.
}
\description{
\code{fread_fwf} takes as basic input a fixed-width filename and its schema and 
wraps around \code{\link[data.table]{fread}} from package \linkS4class{data.table} to provide the
contents of the file.

This method converts the slot \code{Data} from the input \code{StQ} object into a
\linkS4class{data.table} with statistical units by row and variables specified in the input
parameter \code{VarNames} by columns.

To distinguish between variables and qualifiers this function makes use of the slot \code{DD} of
input \linkS4class{StQ} variable.

This method is indeed a wrapper for the function \code{\link[data.table]{dcast.data.table}} of
the package \linkS4class{data.table}, adapted to the structure of object \linkS4class{StQ}.
}
\examples{

}
\seealso{
\code{\link[data.table]{fread}}
}

