% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INExlsxToXML.R
\name{INExlsxToXML}
\alias{INExlsxToXML}
\title{Create an xml file with the contents of an INE xlsx schema file.}
\usage{
INExlsxToXML(xlsxName, sheetToRead = 1, xmlName, regionName = "METADATOS")
}
\arguments{
\item{xlsxName}{Name of the xlsx file containing the schema.}

\item{sheetToRead}{Name or index of the sheet of the xlsx file.}

\item{xmlName}{Name of the xml file is going to be written.}

\item{regionName}{Name of the region in the xlsx file.}
}
\value{
Write the generated xml file.
}
\description{
\code{xlsxToXML} creates an xml file with the contents of the
xlsx file by the Spanish NSI standard for the fwf schema.

The Excel file must have a header in the second row.
}
\examples{
xlsxName    <- file.path(system.file('extdata', package = 'fastReadfwf'), 'dr_EESEadulto_2020.xlsx')
xmlName     <- file.path(system.file('extdata', package = 'fastReadfwf'), 'dr_EESEadulto_2020.xml')
output <- INExlsxToXML(xlsxName = xlsxName, xmlName = xmlName)

}
