% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicroDataSNHS.R
\docType{data}
\name{MicroDataSNHS}
\alias{MicroDataSNHS}
\title{MicroData for the Spanish National Health Survey.}
\format{A \linkS4class{data.table} with 23089 rows (statistical units) and 455 variables

The meaning of each variable is collected in the schema attached as external data to this
package.}
\source{
\url{http://www.ine.es/}
}
\usage{
MicroDataSNHS
}
\description{
A dataset containing the adult variables collected for the Spanish National Health Survey for
2017.
}
\keyword{datasets}
