% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicroDataSNHS.R
\docType{data}
\name{MicroDataSNHS}
\alias{MicroDataSNHS}
\title{MicroData for the Spanish National Health Survey.}
\format{A \linkS4class{data.table} with 5772 rows (statistical units) and 50 variables}
\source{
\url{http://www.ine.es/}
}
\usage{
data(MicroDataSNHS)
}
\description{
\code{MicroDataSNHS} is an object of class \linkS4class{data.table}
containing a subset of the information for adults collected for the Spanish National Health Survey for
2017.
}
