% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StfwfSchema-class.R
\docType{class}
\name{StfwfSchema-class}
\alias{StfwfSchema-class}
\title{S4 class for the standard data.frame for a fixed-width file schema.}
\description{
Definition of the S4 class named \code{StfwfSchema} for the data.frame providing the
schema of the fixed-width file to read.
}
\section{Slots}{

\describe{
\item{\code{df}}{data.frame with 5 columns:
 \itemize{

   \item variable: the name of the variable.
   \item length: the number of positions which the values of this variable occupies in the file.
   \item initialPos: initial position of the field whic hthe values of this variable occupies in
   the file.
   \item finalPos: final position of the field whic hthe values of this variable occupies in the
   file.
   \item type: type of the variable. It must be either \code{num} or \code{char}.
   \item valueRegEx: regular expression for the values of this variable.
   \item description: textual description of the variable.

 }}
}}
\examples{
# An empty StfwfSchema object:
new(Class = 'StfwfSchema')

# A trivial example:
df <- data.frame(variable = c('Turnover', 'Employees'),
                 length = c(9L, 3L),
                 initialPos = c(1, 10),
                 finalPos = c(9, 12),
                 type = rep('num', 2),
                 valueRegEx = c('[0-9]{0,9}', '[0-9]{0,3}'),
                 description = c('Turnover of the business unit',
                                 'Number of employees of the business unit'),
                 stringsAsFactors = FALSE)
new(Class = 'StfwfSchema', df = df)

}

