% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateValues.R
\docType{methods}
\name{validateValues}
\alias{validateValues}
\alias{validateValues,data.frame,StfwfSchema-method}
\title{Fast read a fixed-width file}
\usage{
validateValues(object, StfwfSchema, perl)

\S4method{validateValues}{data.frame,StfwfSchema}(object, StfwfSchema,
  perl = FALSE)
}
\arguments{
\item{object}{Object with rectangular shape whose columns will be validated.}

\item{StfwfSchema}{Object of class \linkS4class{StfwfSchema} with the schema of object.}

\item{perl}{Logical vector of length 1 with default value \code{FALSE} to indicate whether to use
perl or not in the application of regexp.}
}
\value{
Returns \code{TRUE}.
}
\description{
\code{fread_fwf} takes as basic input a fixed-width filename and its schema and
wraps around \code{\link[data.table]{fread}} from package \linkS4class{data.table} to provide the
contents of the file.

This method is indeed a wrapper for the function \code{\link[data.table]{fread}} followed by the
application of \code{\link[stringi]{stri_sub}}.
}
\examples{
\dontrun{
path <- 'C:/Users/David/Documents/Cursos.Seminarios.Impartidos/UCM/EMOS/Organization/Course 2018-2019/Organization - OOP'
pathSchema <- file.path(path, 'EPA/stEPA2018_Schema.xlsx')
stSchema <- XLSToSchema(pathSchema, sheetname = 'stEPA2018_Schema', lang = 'en')
dataFile_T1 <- file.path(path, 'EPA/md_EPA_2018T1.txt')
data_T1_dt <- fread_fwf(dataFile_T1, stSchema, outFormat = 'data.table', perl = TRUE)
data_T1_ti <- fread_fwf(dataFile_T1, stSchema, outFormat = 'tibble', perl = TRUE)
validateValues(data_T1_st, stSchema, perl = TRUE)
validateValues(data_T1_ti, stSchema, perl = TRUE)

}

}

